drive car

+++ libraries/MobileCar/MobileCar.h
@@ -45,7 +45,7 @@ class MobileCar{
 
 			double Dr = 2*pi*radius*rightTicks/TPR;
 			double Dl = 2*pi*radius*leftTicks/TPR;
-			double Dc = (Dr + Dl) / 2;
+			double Dc = (Dr + Dl) / (double)2;
 
 			x = x + Dc * cos(theta);
 			y = y + Dc * sin(theta);
@@ -56,7 +56,7 @@ class MobileCar{
 
 			double velocityToPwm = 255 / MAX_VELOCITY;
 
-			rightPwm = (int)rightPid->execute(Vr_obj*velocityToPwm, Vr_fbk*velocityToPwm);  // radius*Vr_obj*velocityToPwm?
+			rightPwm = (int)rightPid->execute(Vr_obj*velocityToPwm, Vr_fbk*velocityToPwm);  // radius*Vr_obj*velocityToPwm
 			leftPwm = (int)leftPid->execute(Vl_obj*velocityToPwm, Vl_fbk*velocityToPwm);
